requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fake_useragent>=1.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.23.0
openpyxl>=3.1.2
pandas>=2.1.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
selenium>=4.15.0
undetected-chromedriver>=3.5.4
webdriver-manager>=4.0.0
python-dateutil>=2.8.2
gspread>=5.12.0
google-auth>=2.23.0